Move App files in a separate folder
